Move source to src folder
