+++ src/FrozenRandom.js
