+++ src/WeightedRandom.js
